Move Game Logic To Common (#1545)
This PR moves much of the game logic to common. This will make future
performance optimizations likely easier.
