+++ client/src/app/deck/deck-edit.tsx
@@ -3,12 +3,12 @@ import {CARDS, CARDS_LIST} from 'common/cards'
 import {Card, isHermit, isItem} from 'common/cards/types'
 import {EXPANSIONS, ExpansionT} from 'common/const/expansions'
 import {CardEntity, newEntity} from 'common/entities'
+import {getIconPath} from 'common/game/setup-game'
 import {Deck, Tag} from 'common/types/deck'
 import {LocalCardInstance} from 'common/types/server-requests'
 import {sortCardInstances} from 'common/utils/cards'
 import {generateDatabaseCode} from 'common/utils/database-codes'
 import {getCardRank, getDeckCost} from 'common/utils/ranks'
-import {getIconPath} from 'common/utils/state-gen'
 import {validateDeck} from 'common/utils/validation'
 import Accordion from 'components/accordion'
 import Button from 'components/button'
