+++ client/src/app/deck/deck-select.tsx
@@ -1,12 +1,12 @@
 import classNames from 'classnames'
 import {CARDS} from 'common/cards'
+import {getIconPath} from 'common/game/setup-game'
 import {ToastT} from 'common/types/app'
 import {Deck, Tag} from 'common/types/deck'
 import {sortCardInstances} from 'common/utils/cards'
 import {generateDatabaseCode} from 'common/utils/database-codes'
 import {getDeckTypes, sortDecks} from 'common/utils/decks'
 import {getDeckCost} from 'common/utils/ranks'
-import {getIconPath} from 'common/utils/state-gen'
 import {validateDeck} from 'common/utils/validation'
 import Accordion from 'components/accordion'
 import Button from 'components/button'
