+++ client/src/components/game-mode-button/game-mode-button.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames'
 import {CARDS} from 'common/cards'
+import {getIconPath} from 'common/game/setup-game'
 import {ButtonVariant} from 'common/types/buttons'
 import {Deck, Tag} from 'common/types/deck'
 import {getDeckTypes} from 'common/utils/decks'
-import {getIconPath} from 'common/utils/state-gen'
 import Button from 'components/button'
 import Spinner from 'components/spinner'
 import {CopyIcon} from 'components/svgs'
