+++ common/models/game-model.ts
@@ -12,6 +12,11 @@ import query, {ComponentQuery} from '../components/query'
 import {CONFIG, DEBUG_CONFIG} from '../config'
 import {PlayerEntity, SlotEntity} from '../entities'
 import {
+	PlayerSetupDefs,
+	getGameState,
+	setupComponents,
+} from '../game/setup-game'
+import {
 	MultiturnPrimaryAttackDisabledEffect,
 	MultiturnSecondaryAttackDisabledEffect,
 } from '../status-effects/multiturn-attack-disabled'
@@ -43,11 +48,6 @@ import {rowRevive} from '../types/priorities'
 import {PickRequest} from '../types/server-requests'
 import {newIncrementor} from '../utils/game'
 import {newRandomNumberGenerator} from '../utils/random'
-import {
-	PlayerSetupDefs,
-	getGameState,
-	setupComponents,
-} from '../utils/state-gen'
 import {AttackModel, ReadonlyAttackModel} from './attack-model'
 import {BattleLogModel} from './battle-log-model'
