+++ tests/unit/game/advent-of-tcg/hermits/cyberpunkimpulse-rare.test.ts
@@ -14,7 +14,7 @@ import WildItem from 'common/cards/items/wild-common'
 import Efficiency from 'common/cards/single-use/efficiency'
 import {SlotComponent} from 'common/components'
 import query from 'common/components/query'
-import {getAvailableEnergy} from 'server/routines/game'
+import {getAvailableEnergy} from 'common/game/run-game'
 import {testGame} from '../../utils'
 
 describe('Test Cyberpunk Impulse', () => {
